dags/coal_mining_dag.py
import datetime

from airflow.sdk import dag
from airflow.providers.standard.operators.empty import EmptyOperator


@dag(start_date=datetime.datetime(2021, 1, 1), schedule="@daily")
def coal_mining_dag():
    pass


coal_mining_dag()
