requirements.txt
duckdb
clickhouse-driver
numpy
pandas
python-dotenv
mysql-connector-python
dbt-core
dbt-clickhouse